# 10x+ LinkedList performance by eliminating Nil and the abstract tail in Cons